requirements.txt
fastapi
uvicorn
requests
openai
flask
python-dotenv

